+++ pkg-py/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### New features
+
+* New `QueryChat.app()` method enables quicker/easier chatting with a dataset. (#xx)
+* Enabled bookmarking by default in both `.app()` and `.server()` methods. In latter case, you'll need to also specify the `bookmark_store` (either in `shiny.App()` or `shiny.express.app_opts()`) for it to take effect. (#xx)
+
+
+## [UNRELEASED]
+
 ### Changes
 
 * The entire functional API (i.e., `init()`, `sidebar()`, `server()`, etc) has been deprecated in favor of a new class/OOP API. Namely, the new `QueryChat()` class is now the recommended way to start (instead of `init()`), which has methods to replace old functions (e.g., `.sidebar()`, `.server()`, etc). (#101)